Updated logic to only refresh screen after RMC
Found out that each time I refresh the OLED screen, it interferes with the serial reading...
Now we only refresh the screen after speed is updated, which is when the RMC string is processed. 
Only 2 strings are processed, the RCG and RMC. By updating the screen only when the speed is updated, makes sure the RMC string is processed.